refactor(plugin): split service responsibilities
