+++ pkg/plugin/diagnostics.go
@@ -1,0 +1,43 @@
+/*
+Copyright 2025 API Testing Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package plugin
+
+import (
+	"context"
+
+	"github.com/linuxsuren/api-testing/pkg/server"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+)
+
+// PProf returns profiling data for the AI plugin.
+func (s *AIPluginService) PProf(ctx context.Context, req *server.PProfRequest) (*server.PProfData, error) {
+	if err := contextError(ctx); err != nil {
+		return nil, err
+	}
+
+	if req == nil {
+		return nil, status.Error(codes.InvalidArgument, "profiling request cannot be nil")
+	}
+
+	logger := loggerFromContext(ctx)
+	logger.Debug("PProf called", "profile_type", req.Name)
+
+	return &server.PProfData{
+		Data: []byte{},
+	}, nil
+}
